sm-notify: don't rely on non-standard getaddrinfo behavior
getaddrinfo doesn't support both node=NULL and service=NULL,
one or both of these must be specified.

To workaround having to specify one of these, smn_bind_address()
uses getaddrinfo(NULL, "", ...) in the hopes that empty string
counts as specified (non-null) while not actually specifying
the service in terms of limiting the addresses used.
That's my best guess anyway.

Unfortunately this is not valid use of getaddrinfo,
and AFAICT standard says this should result in an error
(especially when specifying AI_NUMERICSERV).

Apparently some platforms don't return error in this case,
although none document this alternate behavior so
hopefully it does whatever NFS hopes it does :).

This invalid usage causes problems when using musl,
which returns an error (as the standards indicate).

From getaddrinfo(3):

-------
If AI_NUMERICSERV  is  specified in  hints.ai_flags  and  service  is not
NULL, then service must point to a string containing a numeric port
number.
-------

The string "" is not a numeric port number.  It is not a number.

And later when describing return values and meaning:

-------
EAI_NONAME
      The node or service is not known; or both node and service
are NULL; or AI_NUMERICSERV was specified in hints.ai_flags and service
was not a numeric port-number string.
-------

Which is what musl's getaddrinfo returns.

As a quick-fix, this uses 'getaddrinfo("0.0.0.0", NULL, ...)"
in this situation which isn't quite right but is enough to fix
sm-notify when built w/musl and using IPV4.

Fixing it "properly" is mostly a question of determining
what NFS "wants" to be done, and reworking code to
avoid relying on non-standard behavior.

There are other instances of this pattern in the codebase,
which should probably be fixing in the same effort.

This article explores this issue a bit on various platforms:
http://klickverbot.at/blog/2012/01/getaddrinfo-edge-case-behavior-on-windows-linux-and-osx/